// src/types/resources/MedicationDispense.ts
/* eslint-disable no-unused-vars */
/* eslint-disable no-trailing-spaces */
// This file is auto-generated by generate_types so do not edit manually

import { TMeta } from '../partials/Meta';
import { TUri } from '../simpleTypes/Uri';
import { TNarrative } from '../partials/Narrative';
import { TResourceContainer } from '../simpleTypes/ResourceContainer';
import { TExtension } from '../partials/Extension';
import { TIdentifier } from '../partials/Identifier';
import { TReference } from '../partials/Reference';
import { TCodeableConcept } from '../partials/CodeableConcept';
import { TMedicationDispensePerformer } from '../partials/MedicationDispensePerformer';
import { TQuantity } from '../partials/Quantity';
import { TDateTime } from '../simpleTypes/DateTime';
import { TAnnotation } from '../partials/Annotation';
import { TDosage } from '../partials/Dosage';
import { TMedicationDispenseSubstitution } from '../partials/MedicationDispenseSubstitution';

export type TMedicationDispense = {
  resourceType?: string;
  id?: string;
  meta?: TMeta;
  implicitRules?: TUri;
  language?: string;
  text?: TNarrative;
  contained?: TResourceContainer[];
  extension?: TExtension[];
  modifierExtension?: TExtension[];
  identifier?: TIdentifier[];
  partOf?: TReference[];
  status: string;
  statusReasonCodeableConcept?: TCodeableConcept;
  statusReasonReference?: TReference;
  category?: TCodeableConcept;
  medicationCodeableConcept?: TCodeableConcept;
  medicationReference?: TReference;
  subject?: TReference;
  context?: TReference;
  supportingInformation?: TReference[];
  performer?: TMedicationDispensePerformer[];
  location?: TReference;
  authorizingPrescription?: TReference[];
  type?: TCodeableConcept;
  quantity?: TQuantity;
  daysSupply?: TQuantity;
  whenPrepared?: TDateTime;
  whenHandedOver?: TDateTime;
  destination?: TReference;
  receiver?: TReference[];
  note?: TAnnotation[];
  dosageInstruction?: TDosage[];
  substitution?: TMedicationDispenseSubstitution;
  detectedIssue?: TReference[];
  eventHistory?: TReference[];
};